webapp/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.1.0
bleach==6.0.0
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.4
django-advance-thumbnail==1.0.0
django-celery-beat==2.5.0
django-constance==3.1.0
django-picklefield==3.1
django-rest-framework==0.1.0
django-summernote==0.8.20.0
django-timezone-field==6.0
djangorestframework==3.14.0
kombu==5.3.1
Pillow==10.0.0
prompt-toolkit==3.0.39
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
